fixed warnings

+++ common/global.c
@@ -1457,7 +1457,7 @@ void debugDumpMemory(const void *address, uint length, bool printAddress)
   while (z < length)
   {
     p = (const byte*)address+z;
-    if (printAddress) fprintf(stderr,"%08lx:",(unsigned long)p);
+    if (printAddress) fprintf(stderr,"%08lx:",(intptr_t)p);
     fprintf(stderr,"%08lx  ",(unsigned long)(p-(byte*)address));
 
     for (j = 0; j < 16; j++)
